# Optimize spatial predicates when vectors are constant.