--- a/spatial/src/spatial/geos/functions/scalar/st_covered_by.cpp
+++ b/spatial/src/spatial/geos/functions/scalar/st_covered_by.cpp
@@ -3,6 +3,7 @@
 #include "spatial/geos/functions/scalar.hpp"
 #include "spatial/geos/functions/common.hpp"
 #include "spatial/geos/geos_wrappers.hpp"
+#include "spatial/geos/geos_executor.hpp"
 
 #include "duckdb/parser/parsed_data/create_scalar_function_info.hpp"
 #include "duckdb/common/vector_operations/unary_executor.hpp"
@@ -16,14 +17,10 @@ using namespace spatial::core;
 
 static void CoveredByFunction(DataChunk &args, ExpressionState &state, Vector &result) {
 	auto &lstate = GEOSFunctionLocalState::ResetAndGet(state);
-
-	BinaryExecutor::Execute<string_t, string_t, bool>(args.data[0], args.data[1], result, args.size(), [&](string_t &left_blob, string_t &right_blob) {
-		auto left_geometry = lstate.factory.Deserialize(left_blob);
-		auto right_geometry = lstate.factory.Deserialize(right_blob);
-		auto geos_left = lstate.ctx.FromGeometry(left_geometry);
-		auto geos_right = lstate.ctx.FromGeometry(right_geometry);
-		return geos_left.CoveredBy(geos_right);
-	});
+	auto &left = args.data[0];
+	auto &right = args.data[1];
+	auto count = args.size();
+	GEOSExecutor::ExecuteNonSymmetricPreparedBinary(lstate, left, right, count, result, GEOSCoveredBy_r, GEOSPreparedCoveredBy_r);
 }
 
 void GEOSScalarFunctions::RegisterStCoveredBy(ClientContext &context) {