… README.md
… spatial/src/spatial/geos/functions/scalar/st_touches.cpp
… spatial/src/spatial/geos/functions/scalar/st_within.cpp
… spatial/src/spatial/geos/geos_wrappers.cpp