### Misc
- We now properly throw exceptions on GEOS errors.

### Functions & Casts
- Added `ST_ContainsProperly`

### Optimizations
We now create GEOS "prepared geometries" when executing spatial predicates when we know that one of the arguments is a DuckDB ConstantVector. We also optimize `ST_Distance` and `ST_DistanceWithin`. Being able to do this optimization is a really cool way to leverage how DuckDB's vectorized execution engine works.

Originally I thought that this would only be a performance win in the case when you want to apply a filter with a specific constant geometry, like in `SELECT geom WHERE ST_Intersects(geom, ST_GeomFromText('...')) FROM tbl` but it turns out that we also produce ConstantVectors when e.g. doing joins, which gave spatial joins a surprisingly big boost. We still don't have any general spatial indexes though so It's still not fast enough to join all 1000000 rides with all 264 zones in our taxi dataset in real-time, but here's a micro benchmark.

Setup:
```sql
load parquet;
CREATE TABLE rides AS SELECT * FROM 'spatial/test/data/nyc_taxi/yellow_tripdata_2010-01-limit1mil.parquet';
CREATE TABLE zones AS SELECT * FROM st_read('./spatial/test/data/nyc_taxi/taxi_zones/');
PRAGMA enable_profiling;
```

Running this query:
```sql
 SELECT rides.pickup_latitude, rides.pickup_longitude, z.zone FROM (SELECT * FROM zones LIMIT 1) as z JOIN rides ON ST_Contains(st_geomfromwkb(z.wkb_geometry), ST_Transform(st_point(pickup_latitude, pickup_longitude), 'EPSG:4326', 'ESRI:102718'));
 ```
 As we increase the number of zones, 'n' in `... LIMIT n) as z`:
 | n | before | after |
 |--|-------|-------|
 |1 |  2.43s   | 1.46s |
 |2|  25.56s  | 2.95s |
 |3| 28.26s   | 4.40s |
 |10| 43.83s  | 14.26s |
 |20| 68.27 | 28.56 |
 
 ![chart](https://user-images.githubusercontent.com/1337636/235321042-983db782-5f82-4bd0-9a96-d555bd933b1d.png)